pyproject.toml
[tool.poetry]
name = "aws-glue-etl-extensions"
version = "0.1.0"
description = ""
authors = ["Martin Macak <martin.macak@gmail.com>"]
readme = "README.md"
packages = [
    { include = "aws-glue-etl-extensions", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10,<3.11"

[tool.poetry.group.dev.dependencies]
boto3 = "^1.34.59"
pyspark = "^3.2.0"
jupyter = "^1.0.0"
aws-glue-sessions = "^1.0.4"
awsglue3-local = "^1.0.0"

[tool.poetry-dynamic-versioning]
enable = true
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
